Clean up imports and reorganize code
- Removed unused imports related to Lattice, SpaceGroup, Phase, and Site.
- Streamlined import statements across multiple files for better readability.
- Ensured all necessary components are still imported correctly.
