+++ src/easydiffraction/__init__.py
@@ -2,14 +2,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # © 2021-2024 Contributors to the EasyDiffraction project <https://github.com/EasyScience/EasyDiffraction>
 
-from easycrystallography.Components.Lattice import Lattice as Lattice
-from easycrystallography.Components.SpaceGroup import SpaceGroup as SpaceGroup
-
 from .job.job import DiffractionJob as Job
-from .job.model.phase import Phase as Phase
-from .job.model.phase import Phases as Phases
-from .job.model.site import Atoms as Atoms
-from .job.model.site import Site as Site
 from .utils import download_from_repository
 
 __all__ = ['Job', 'download_from_repository']
