+++ src/easydiffraction/calculators/pdffit2/wrapper.py
@@ -7,17 +7,17 @@ from typing import List
 
 import numpy as np
 from diffpy.utils.parsers.loaddata import loadData
+from easycrystallography.Components.Lattice import Lattice
+from easycrystallography.Components.SpaceGroup import SpaceGroup
 from easyscience import global_object as borg
 from easyscience.Objects.Inferface import ItemContainer
 
-from easydiffraction import Lattice
-from easydiffraction import Phases
-from easydiffraction import Site
-from easydiffraction import SpaceGroup
 from easydiffraction.calculators.pdffit2.calculator import Pdffit2 as Pdffit2_calc
 from easydiffraction.calculators.wrapper_base import WrapperBase
 from easydiffraction.job.experiment.pd_1d import PDFParameters
 from easydiffraction.job.experiment.pd_1d import Powder1DParameters
+from easydiffraction.job.model.phase import Phases
+from easydiffraction.job.model.site import Site
 
 if TYPE_CHECKING:
     pass
