+++ src/easydiffraction/job/old_sample/old_sample.py
@@ -12,8 +12,6 @@ from easyscience.Datasets.xarray import xr
 from easyscience.global_object.undo_redo import property_stack_deco
 from easyscience.Objects.ObjectClasses import BaseObj
 
-from easydiffraction import Phase
-from easydiffraction import Phases
 from easydiffraction.calculators.wrapper_factory import WrapperFactory
 from easydiffraction.calculators.wrapper_types import Neutron
 from easydiffraction.calculators.wrapper_types import Powder
@@ -21,6 +19,8 @@ from easydiffraction.job.experiment.pd_1d import Instrument1DCWParameters
 from easydiffraction.job.experiment.pd_1d import Instrument1DTOFParameters
 from easydiffraction.job.experiment.pd_1d import PolPowder1DParameters as Pattern1D_Pol
 from easydiffraction.job.experiment.pd_1d import Powder1DParameters as Pattern1D
+from easydiffraction.job.model.phase import Phase
+from easydiffraction.job.model.phase import Phases
 
 
 class Sample(BaseObj):
